Add a remove_entry function
Removing/modifying entries is not yet possible with this ini lib,
but I did need it for a project it's used in.

The function I added frees the entry-to-be-removed's key and value,
and then uses memmove to shift all entries after the removed entry
forward. Finally, it reduces the section size by 1.